Refactor hooks import paths; add OptimisticFn to mutation utils
